Update 008. Merge two sorted arrays
You are given two integer arrays nums1 and nums2 arranged in non-decreasing order, and two integers m and n, representing the number of elements in nums1 and nums2 respectively.

Please merge nums2 into nums1 so that the merged array is also arranged in non-decreasing order.

Note: Ultimately, the merged array should not be returned by the function, but stored in the array nums1. To cope with this situation, nums1 has an initial length of m + n, where the first m elements represent elements that should be merged and the last n elements are 0 and should be ignored. The length of nums2 is n .